# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tempo API",
    "version": "0.5.0"
  },
  "paths": {
...
    "/api/exercises": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "List Exercises",
        "operationId": "list_exercises_api_exercises_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Free-text match on name",
              "title": "Q"
            },
            "description": "Free-text match on name"
          },
          {
            "name": "muscle",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Muscle"
            }
          },
          {
            "name": "equipment",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Equipment"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseListOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Create Exercise",
        "operationId": "create_exercise_api_exercises_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/by-slug/{slug}": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "Get Exercise By Slug",
        "description": "Resolve a url-safe slug to its exercise (the web Library's detail-page lookup).",
        "operationId": "get_exercise_by_slug_api_exercises_by_slug__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/{exercise_id}": {
      "get": {
        "tags": [
          "exercises"
        ],
        "summary": "Get Exercise",
        "operationId": "get_exercise_api_exercises__exercise_id__get",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "exercises"
        ],
        "summary": "Update Exercise",
        "description": "Edit a **custom** exercise. Renaming regenerates the slug, keeping the old one as an alias.",
        "operationId": "update_exercise_api_exercises__exercise_id__patch",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExerciseUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "exercises"
        ],
        "summary": "Delete Exercise",
        "description": "Soft-delete a custom exercise, refusing to orphan the sets that reference it.",
        "operationId": "delete_exercise_api_exercises__exercise_id__delete",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          },
          {
            "name": "reassign_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Reassign To"
            }
          },
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Dry Run"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDeleteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/exercises/{exercise_id}/illustration": {
      "post": {
        "tags": [
          "exercises"
        ],
        "summary": "Ensure Illustration",
        "description": "On-demand fallback (docs/06): generate + persist this exercise's illustration if missing.\n\nReturns the exercise; ``illustration_status`` is ``ready`` (with ``illustration_url``) on\nsuccess, or ``generating`` if a batch run is already producing it. 503 if the image\nprovider/Blob is unavailable.",
        "operationId": "ensure_illustration_api_exercises__exercise_id__illustration_post",
        "parameters": [
          {
            "name": "exercise_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Exercise Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExerciseDetailOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExerciseCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Name"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "force": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Force"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "mechanic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mechanic"
          },
          "equipment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment"
          },
          "primary_muscles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Primary Muscles"
          },
          "secondary_muscles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Secondary Muscles"
          },
          "instructions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Instructions"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ExerciseCreate",
        "description": "Create a custom exercise. ``slug`` is derived from ``name`` when omitted."
      },
      "ExerciseDeleteOut": {
        "properties": {
          "exercise": {
            "$ref": "#/components/schemas/ExerciseOut"
          },
          "set_count": {
            "type": "integer",
            "title": "Set Count"
          },
          "reassigned_to": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExerciseOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run"
          }
        },
        "type": "object",
        "required": [
          "exercise",
          "set_count",
          "reassigned_to",
          "dry_run"
        ],
        "title": "ExerciseDeleteOut",
        "description": "What a custom-exercise delete did — or, with ``dry_run``, would have done."
      },
      "ExerciseDetailOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "force": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Force"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "mechanic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mechanic"
          },
          "equipment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment"
          },
          "primary_muscles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Primary Muscles"
          },
          "secondary_muscles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Secondary Muscles"
          },
          "illustration_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Illustration Url"
          },
          "illustration_url_light": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Illustration Url Light"
          },
          "illustration_status": {
            "type": "string",
            "title": "Illustration Status"
          },
          "is_custom": {
            "type": "boolean",
            "title": "Is Custom",
            "default": false
          },
          "instructions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Instructions"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "category",
          "force",
          "level",
          "mechanic",
          "equipment",
          "primary_muscles",
          "secondary_muscles",
          "illustration_url",
          "illustration_url_light",
          "illustration_status",
          "instructions",
          "source",
          "created_at",
          "updated_at"
        ],
        "title": "ExerciseDetailOut",
        "description": "Full detail: adds the how-to steps + provenance/timestamps."
      },
      "ExerciseListOut": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExerciseOut"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "title": "ExerciseListOut"
      },
      "ExerciseOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "force": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Force"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "mechanic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mechanic"
          },
          "equipment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment"
          },
          "primary_muscles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Primary Muscles"
          },
          "secondary_muscles": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Secondary Muscles"
          },
          "illustration_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Illustration Url"
          },
          "illustration_url_light": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Illustration Url Light"
          },
          "illustration_status": {
            "type": "string",
            "title": "Illustration Status"
          },
          "is_custom": {
            "type": "boolean",
            "title": "Is Custom",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "slug",
          "name",
          "category",
          "force",
          "level",
          "mechanic",
          "equipment",
          "primary_muscles",
          "secondary_muscles",
          "illustration_url",
          "illustration_url_light",
          "illustration_status"
        ],
        "title": "ExerciseOut",
        "description": "Catalog summary (list rows + nested group headers)."
      },
...
      "ExerciseUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 200,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "force": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Force"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level"
          },
          "mechanic": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mechanic"
          },
          "equipment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Equipment"
          },
          "primary_muscles": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Muscles"
          },
          "secondary_muscles": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secondary Muscles"
          },
          "instructions": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instructions"
          }
        },
        "type": "object",
        "title": "ExerciseUpdate",
        "description": "Patch a **custom** exercise; only supplied fields change.\n\nNo ``slug``: it is derived from ``name``, so the two cannot drift. Renaming keeps the old slug\nresolvable as an alias."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}